+++ pipelines/model_qwen.py
@@ -25,6 +25,8 @@ def load_qwen(checkpoint_info, diffusers_load_config=None):
         diffusers.pipelines.auto_pipeline.AUTO_TEXT2IMAGE_PIPELINES_MAPPING["qwen-image"] = diffusers.QwenImageEditPipeline
         diffusers.pipelines.auto_pipeline.AUTO_IMAGE2IMAGE_PIPELINES_MAPPING["qwen-image"] = diffusers.QwenImageEditPipeline
         diffusers.pipelines.auto_pipeline.AUTO_INPAINT_PIPELINES_MAPPING["qwen-image"] = diffusers.QwenImageEditPipeline
+    elif 'Layered' in repo_id:
+        cls_name = diffusers.QwenImageLayeredPipeline
     else:
         cls_name = diffusers.QwenImagePipeline
         diffusers.pipelines.auto_pipeline.AUTO_TEXT2IMAGE_PIPELINES_MAPPING["qwen-image"] = diffusers.QwenImagePipeline
@@ -69,6 +71,10 @@ def load_qwen(checkpoint_info, diffusers_load_config=None):
     pipe.task_args = {
         'output_type': 'np',
     }
+    if 'Layered' in repo_id:
+        pipe.task_args['use_en_prompt'] = True
+        pipe.task_args['cfg_normalize'] = True
+        pipe.task_args['layers'] = shared.opts.model_qwen_layers
 
     del text_encoder
     del transformer